requirements.txt
Django==1.11.10
django-cors-headers==2.1.0
django-filter==1.1.0
djangorestframework==3.7.7
numpy==1.14.0
pytz==2018.3
